fix: corrige e reorganiza testes
