+++ __tests__/unit/lista-nominal/table/formatters/time.test.ts
@@ -1,4 +1,4 @@
-import { formatDate, isDate, parseDate } from "@/common/time";
+import { formatDate, isDate, parseDate } from "@/features/common/time";
 
 describe("formatDate", () => {
     it("deve formatar uma data corretamente no formato dd/mm/aa", () => {
